# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CourierStatus": {
        "enum": [
          "not_dispatched",
          "in_transit",
          "delivered",
          "returned",
          "lost"
        ],
        "title": "CourierStatus",
        "type": "string"
      },
...
      "CreateOrderRequest": {
        "properties": {
          "amount_bdt": {
            "exclusiveMinimum": 0.0,
            "maximum": 200000.0,
            "title": "Amount Bdt",
            "type": "integer"
          },
          "buyer_id": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Buyer Id",
            "type": "string"
          },
          "product_category": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Product Category",
            "type": "string"
          },
          "seller_id": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Seller Id",
            "type": "string"
          }
        },
        "required": [
          "buyer_id",
          "seller_id",
          "amount_bdt",
          "product_category"
        ],
        "title": "CreateOrderRequest",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorOut": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "LedgerAccount": {
        "enum": [
          "BUYER_WALLET",
          "HOLD",
          "SELLER_WALLET",
          "REFUND"
        ],
        "title": "LedgerAccount",
        "type": "string"
      },
      "LedgerEntryOut": {
        "properties": {
          "account": {
            "$ref": "#/components/schemas/LedgerAccount"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "credit": {
            "title": "Credit",
            "type": "integer"
          },
          "debit": {
            "title": "Debit",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "account",
          "debit",
          "credit",
          "reason",
          "created_at"
        ],
        "title": "LedgerEntryOut",
        "type": "object"
      },
...
      "OrderOut": {
        "properties": {
          "amount_bdt": {
            "title": "Amount Bdt",
            "type": "integer"
          },
          "buyer_id": {
            "title": "Buyer Id",
            "type": "string"
          },
          "can_report_problem": {
            "title": "Can Report Problem",
            "type": "boolean"
          },
          "courier_status": {
            "$ref": "#/components/schemas/CourierStatus"
          },
          "delivered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          },
          "delivery_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Sandbox only: returned once, when the order is created.",
            "title": "Delivery Code"
          },
          "dispute_ids": {
            "items": {
              "type": "string"
            },
            "title": "Dispute Ids",
            "type": "array"
          },
          "hold_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hold Until"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "ledger": {
            "items": {
              "$ref": "#/components/schemas/LedgerEntryOut"
            },
            "title": "Ledger",
            "type": "array"
          },
          "ledger_balanced": {
            "title": "Ledger Balanced",
            "type": "boolean"
          },
          "placed_at": {
            "format": "date-time",
            "title": "Placed At",
            "type": "string"
          },
          "product_category": {
            "title": "Product Category",
            "type": "string"
          },
          "released_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Released At"
          },
          "seller_id": {
            "title": "Seller Id",
            "type": "string"
          },
          "server_time": {
            "title": "Server Time",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEventOut"
            },
            "title": "Timeline",
            "type": "array"
          }
        },
        "required": [
          "id",
          "buyer_id",
          "seller_id",
          "product_category",
          "amount_bdt",
          "status",
          "placed_at",
          "delivered_at",
          "hold_until",
          "released_at",
          "courier_status",
          "ledger",
          "ledger_balanced",
          "timeline",
          "dispute_ids",
          "can_report_problem",
          "server_time"
        ],
        "title": "OrderOut",
        "type": "object"
      },
      "OrderStatus": {
        "enum": [
          "HELD",
          "DELIVERED",
          "DISPUTABLE",
          "DISPUTED",
          "RELEASED",
          "REFUNDED",
          "ESCALATED"
        ],
        "title": "OrderStatus",
        "type": "string"
      },
...
      "TimelineEventOut": {
        "properties": {
          "event": {
            "title": "Event",
            "type": "string"
          },
          "t": {
            "title": "T",
            "type": "string"
          }
        },
        "required": [
          "t",
          "event"
        ],
        "title": "TimelineEventOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SafeOrder",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/orders": {
      "post": {
        "description": "Create a Safe Order and hold the money (simulated). The sandbox code is returned once.",
        "operationId": "create_order_api_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad request, for example a wrong delivery code"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A model is not available"
          }
        },
        "summary": "Create Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{order_id}": {
      "get": {
        "operationId": "get_order_api_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad request, for example a wrong delivery code"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A model is not available"
          }
        },
        "summary": "Get Order",
        "tags": [
          "orders"
        ]
      }
    },
...
  }
}